% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{add}
\alias{add}
\title{Add two numbers}
\usage{
add(x, y, z = 0)
}
\arguments{
\item{x}{first number}

\item{y}{second number}

\item{z}{third number}
}
\value{
the result of addition
}
\description{
Add two numbers
}
