% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-pipeline_core.R
\name{append_fit_results}
\alias{append_fit_results}
\title{Append fit results}
\usage{
append_fit_results(data, linear_threshold)
}
\arguments{
\item{data}{A tibble containing multicomponent fluorescence 'fam' and cycle number 'cycle' data
separated by well_position and batch_name}

\item{linear_threshold}{A numeric defining the minimal overall change-in-fluorescence required to
attempt to optimize to a 5-parameter logistic regression}
}
\value{
A tibble
}
\description{
This function separately calls the run_fit_curve() function on all reactions contained
in a tibble and appends the model curve parameters and features to the data set. This
workflow enables batch computation for model curves optimized to the qPCR data.
}
\examples{
data <- tidy_run_data |> dplyr::filter(target_name == "S. pneumoniae_Ba06439619_s1")
curve_fit_data <- append_fit_results(data = data, linear_threshold = 400)
}
