% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_calls_with_key.R
\name{assign_calls_with_key}
\alias{assign_calls_with_key}
\title{Assign calls with key}
\usage{
assign_calls_with_key(data, key_path)
}
\arguments{
\item{data}{A tibble containing qPCR and model curve data for resulting}

\item{key_path}{file path to the target-threshold key associating thresholds
for each assay target}
}
\value{
A tibble
}
\description{
Assigns positive and negative calls to each reaction depending on how the features
of the model regression compare to the thresholds defined for the respective assay target.
Appends these calls to the data frame and returns the modified tibble.
}
\examples{
key_path = system.file("extdata", "target_threshold_key.xlsx", package = "OAtools")
results <- assign_calls_with_key(data = curve_fit_data, key_path = key_path)
}
