% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignCallsToSE.R
\name{assignCallsToSE}
\alias{assignCallsToSE}
\title{Assign Final Results to Each Well}
\usage{
assignCallsToSE(se, key_path)
}
\arguments{
\item{se}{a SummarizedExperiment object containing OpenArray qPCR Data}

\item{key_path}{file path to the target-threshold key}
}
\value{
a SummarizedExperiment
}
\description{
Assigns positive or negative results to every PCR reaction depending on the
features of the model curve and the thresholds defined for the relevant
gene. These calls are appended to the colData of the SummarizedExperiment.
}
\examples{
data(example_se)
key_path = system.file(
    "extdata", 
    "target_threshold_key.xlsx", 
    package = "OAtools"
)

se <- example_se |> 
    fitModelsToSE(linear_threshold = 500) |> 
    assignCallsToSE(key_path = key_path)
}
