% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeModels.R
\name{computeModels}
\alias{computeModels}
\title{Optimize 5PL models to OpenArray PCR data and save as metadata}
\usage{
computeModels(se, assay_name, linear_threshold = 400)
}
\arguments{
\item{se}{OpenArray PCR data as a SummarizedExperiment object}

\item{assay_name}{character value specifying the assay matrix from which to
load PCR data for curve-fitting, either \code{fluo_normalized} or \code{fluo_reporter}}

\item{linear_threshold}{numeric value specifying the minimum overall
change-in-fluorescence over the PCR reaction required for the optimizer
to attempt fitting a logistic model}
}
\value{
OpenArray PCR data as a SummarizedExperiment object with
information from the computed model stored as metadata.
}
\description{
Computes 5-parameter logistic models optimized to PCR data contained in the
specified assay matrix, either \code{fluo_normalized} or \code{fluo_reporter}. The
former refers to the base-line adjusted, normalized fluorescence from the
\verb{Amplification Data} tab of the original Excel output. The latter refers to
the multicomponent fluorescence from the \verb{Multicomponent Data} tab.

The expected input for this function is a SummarizedExperiment object
containing OpenArray PCR data, which can be generated by calling the
\code{excelToSE()} function of the package on the raw Excel output from
QuantStudio software.
}
\details{
Under the hood, this function invokes .runFitCurve(), a thin wrapper which
calls python3 code to fit models with scipy.optimize, on each PCR reaction
of the OpenArray plate separately. The computed models are added to the
SummarizedExperiment container in metadata, named as \code{assay_name} +
\verb{_models}. For example, \code{fluo_normalized_models} would be created in the
experiment metadata if \code{computeModels(assay_name = "fluo_normalized")} is
called.
}
\examples{
path <- system.file(
    "extdata", 
    "oa_gene_expression_1.xlsx", 
    package = "OAtools"
)

se <- excelToSE(excel_path = path) |> 
    computeModels(assay_name = "fluo_normalized") |> 
    computeModels(assay_name = "fluo_reporter")
}
