% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determinePCRResults.R
\name{determinePCRResults}
\alias{determinePCRResults}
\title{Determine PCR results from fit curves}
\usage{
determinePCRResults(se, key_path)
}
\arguments{
\item{se}{a SummarizedExperiment object containing OpenArray qPCR Data}

\item{key_path}{file path to the target-threshold key}
}
\value{
a SummarizedExperiment
}
\description{
Assigns positive or negative PCR results to each reaction depending on the
equation of the model optimized to observed multicomponent fluorescence
values. The reporter dye fluorescence without normalization is used to
distinguish between real amplification and false positives.
}
\details{
The key is an Excel file storing values used by \code{determinePCRResults()} to
categorize PCR curves as \code{positive} or \code{negative}. In the key, each gene is
associated with threshold values for Crt, change-in-fluorescence, and slope
at the reaction midpoint.

For more specifics, input \code{?target_threshold_key} into the R console.
}
\examples{
data(example_se)

key_path = system.file(
    "extdata", 
    "target_threshold_key.xlsx", 
    package = "OAtools"
)

se <- example_se |> 
    determinePCRResults(key_path = key_path)
}
