% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildApp.R
\name{.buildServer}
\alias{.buildServer}
\title{Web Server}
\usage{
.buildServer(input, output, session)
}
\arguments{
\item{input}{reactive values from UI widgets}

\item{output}{render functions for results}

\item{session}{environment for user session}
}
\value{
A Shiny web server
}
\description{
Helper method to buildApp() responsible for construction
of the Shiny web server
}
