% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildDataImport.R
\name{.serverDataImport}
\alias{.serverDataImport}
\title{Data Import Server}
\usage{
.serverDataImport(id)
}
\arguments{
\item{id}{identifier unique to \verb{Data Import} page}
}
\value{
The \verb{Data Import} server component
}
\description{
Data Import Server
}
