% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildFitCurves.R
\name{.serverFitCurves}
\alias{.serverFitCurves}
\title{Curve Fitting Server}
\usage{
.serverFitCurves(id, se)
}
\arguments{
\item{id}{identifier unique to \verb{Fit Curves} page}

\item{se}{SummarizedExperiment with OpenArray assay data}
}
\value{
The \verb{Fit Curves} server component
}
\description{
Curve Fitting Server
}
