% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildGraphics.R
\name{.serverGraphics}
\alias{.serverGraphics}
\title{Graphics Server}
\usage{
.serverGraphics(id, se)
}
\arguments{
\item{id}{identifier unique to \code{Graphics} page}

\item{se}{SummarizedExperiment with OpenArray assay data}
}
\value{
The \code{Graphics} server component
}
\description{
Graphics Server
}
