% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildDataImport.R
\name{.uiDataImport}
\alias{.uiDataImport}
\title{Data Import UI}
\usage{
.uiDataImport(id)
}
\arguments{
\item{id}{identifier unique to \verb{Data Import} page}
}
\value{
The \verb{Data Import} page
}
\description{
Data Import UI
}
