% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildFitCurves.R
\name{.uiFitCurves}
\alias{.uiFitCurves}
\title{Curve Fitting UI}
\usage{
.uiFitCurves(id)
}
\arguments{
\item{id}{identifier unique to \verb{Fit Curves} page}
}
\value{
The \verb{Fit Curves} page
}
\description{
Curve Fitting UI
}
