% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildGraphics.R
\name{.uiGraphics}
\alias{.uiGraphics}
\title{Graphics UI}
\usage{
.uiGraphics(id)
}
\arguments{
\item{id}{identifier unique to \code{Graphics} page}
}
\value{
The \code{Graphics} page
}
\description{
Graphics UI
}
