% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildReporting.R
\name{.uiReporting}
\alias{.uiReporting}
\title{Reporting UI}
\usage{
.uiReporting(id)
}
\arguments{
\item{id}{identifier unique to \code{Reporting} page}
}
\value{
The \code{Reporting} page
}
\description{
Reporting UI
}
