% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excelToSE.R
\name{excelToSE}
\alias{excelToSE}
\title{Convert OpenArray data from Excel to a Summarized Experiment}
\usage{
excelToSE(excel_path, header_rows = 17, skip = 19)
}
\arguments{
\item{excel_path}{file path to the Excel document containing the PCR data}

\item{header_rows}{number of rows of run metadata to read in from the header}

\item{skip}{number of rows to skip when reading fluorescence data or results}
}
\value{
OpenArray PCR data as a SummarizedExperiment object
}
\description{
Transforms raw gene expression run data exported from the OpenArray
QuantStudio 12K Flex Software from .xlsx format into an instance of the
SummarizedExperiment class from Bioconductor.
}
\examples{
path = system.file(
    "extdata", 
    "oa_gene_expression_1.xlsx", 
    package = "OAtools"
)

se <- excelToSE(excel_path = path)
}
