% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excelToSummarizedExperiment.R
\name{excelToSummarizedExperiment}
\alias{excelToSummarizedExperiment}
\title{Convert OpenArray data from Excel to a Summarized Experiment}
\usage{
excelToSummarizedExperiment(path = path, num_results = 96)
}
\arguments{
\item{path}{file path to the run data}

\item{num_results}{an integer representing the number of rows on the
'results' sheet to be imported.}
}
\value{
A SummarizedExperiment object
}
\description{
Transforms raw gene expression run data exported from the OpenArray
QuantStudio 12K Flex Software from .xlsx format into an instance of the
SummarizedExperiment class from Bioconductor.
}
\examples{
path = system.file(
    "extdata", 
    "oa_gene_expression_batch1.xlsx", 
    package = "OAtools"
)
se <- excelToSummarizedExperiment(path = path, num_results = 96)
}
