% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{find_gaps}
\alias{find_gaps}
\title{Title}
\usage{
find_gaps(df, r_sst_mean, r_sst_range, r_do_mean, r_do_range,
  fld_lon = "Longitude", fld_lat = "Latitude", distanceweight = 10^-11,
  temporalweight = 10)
}
\arguments{
\item{df}{data frame subsetted to OA rows}

\item{fld_lon}{name of field with longitude in df}

\item{fld_lat}{name of field with latitude in df}

\item{temporalweight}{}
}
\description{
Title
}
