% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitModelsToSE.R
\name{fitModelsToSE}
\alias{fitModelsToSE}
\title{Fit Models to SummarizedExperiment}
\usage{
fitModelsToSE(se, linear_threshold)
}
\arguments{
\item{se}{A SummarizedExperiment object with OpenArray fluorescence data}

\item{linear_threshold}{a numeric defining the minimal
change-in-fluorescence to fit a logistic model.}
}
\value{
a SummarizedExperiment
}
\description{
This function invokes the run_fit_curve() function for each PCR reaction
contained in the given SummarizedExperiment object, adding parameters and
features of the curves to the dataset. This function expects as input the
result of the excel_to_summarized_experiment() function of the package.
}
\details{
PCR reactions with a change-in-fluorescence exceeding the linear_threshold
are fit to a logistic regression. To decrease computing time, a curves that
fail to meet this threshold are considered trivially negative and fit
instead to a linear model.
}
\examples{
path = system.file(
    "extdata", 
    "oa_gene_expression_1.xlsx", 
    package = "OAtools"
)

se <- excelToSummarizedExperiment(path = path, num_results = 96)
se <- fitModelsToSE(se = se, linear_threshold = 500)
}
