% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-pipeline_core.R
\name{format_results}
\alias{format_results}
\title{Format results for reporting}
\usage{
format_results(data, include_fluorescence_data)
}
\arguments{
\item{data}{A tibble containing analyzed qPCR data}

\item{include_fluorescence_data}{logical determining whether fluorescence vs. cycle data is included}
}
\value{
A reformatted tibble for reporting
}
\description{
Reformats analyzed OpenArray qPCR data, converting the long tidy tibble into a wider,
human-readable one in which each row reflects one reaction. Fluorescence data for each
cycle may be optionally included.
}
\examples{
formatted_data <- format_results(data = result_data, include_fluorescence_data = TRUE)
}
