% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_expression_run_data.R
\name{gene_expression_run_data}
\alias{gene_expression_run_data}
\title{Gene Expression Run Data}
\format{
An Excel file (\code{.xlsx}) with two tabs:
\describe{
\item{Multicomponent Data}{Multicomponent fluorescence values for each well position and cycle number}
\item{Results}{Run results and QC metrics output by QuantStudio Software}
}
}
\source{
Exported from QuantStudio 12K Flex Software after a gene expression run, then filtered down to a smaller file size.
}
\description{
This Excel file (.xlsx) contains an example subset of OpenArray gene expression run data intended for use with package examples and unit testing. This dataset includes twelve specimens tested on four targets.
}
\details{
Example run data is stored in the \verb{inst/extdata/} directory of the package. Access this file with:

\code{system.file("extdata", "gene_expression_run_data.xlsx", package = "OAtools")}.
}
\keyword{data}
