% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{generate_report_from_se}
\alias{generate_report_from_se}
\title{Generate a report}
\usage{
generate_report_from_se(se, path)
}
\arguments{
\item{se}{a SummarizedExperiment containing OpenArray qPCR data}

\item{path}{intended outfile path}
}
\value{
an .html report summarizing the OpenArray run results
}
\description{
Generates a .html report summarizing the OpenArray experiment.
}
