% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_rundata.R
\docType{data}
\name{oa_gene_expression_1}
\alias{oa_gene_expression_1}
\title{Example Raw OpenArray Gene Expression Data}
\format{
An Excel file (\code{.xlsx}) with three tabs:
\describe{
\item{Amplification Data}{normalized fluorescence by cycle number}
\item{Multicomponent Data}{spectral contribution of the reporter dye
by cycle number}
\item{Results}{PCR results, sample metadata, and QC metrics}
}
}
\source{
This file was exported from QuantStudio 12K Flex Software after a gene
expression run, then filtered down to include 12 samples and 4 genes for
file size concerns.
}
\description{
The first of two Excel files storing run data from separate OpenArray
gene expression experiments. The context behind the experiment was
respiratory tract microbiota profiling on human nasopharyngeal swabs from
patients experiencing respiratory syndromes.
}
\details{
Access this file with:
\code{system.file("extdata", "oa_gene_expression_1.xlsx", package = "OAtools")}
}
\keyword{data}
