% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-visualizations.R
\name{plot_crt_from_se}
\alias{plot_crt_from_se}
\title{Plot Crt Values from SummarizedExperiment}
\usage{
plot_crt_from_se(se)
}
\arguments{
\item{se}{a SummarizedExperiment object containing OpenArray qPCR data}
}
\value{
a ggplot2 figure
}
\description{
Generates a box and whisker plot to visualize the distribution of observed
Crt values for each assay target in an OpenArray plate.
}
