% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_overview.R
\name{plot_overview}
\alias{plot_overview}
\title{Plot Overview of Amplifications as a Heat Map}
\usage{
plot_overview(se)
}
\arguments{
\item{se}{a SummarizedExperiment object containing OpenArray qPCR data}
}
\value{
a ggplot2 figure
}
\description{
This function generates a heat map summarizing the amplification status
results for each sample on the OpenArray plate by target gene.
}
\examples{
data(example_se)
plot_overview(example_se)
}
