% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_qc.R
\name{plot_qc}
\alias{plot_qc}
\title{Plot a 3D Quality Control Graphic from a SummarizedExperiment}
\usage{
plot_qc(se)
}
\arguments{
\item{se}{a SummarizedExperiment object containing OpenArray qPCR data}
}
\value{
a plotly figure
}
\description{
Generates a 3-dimensional quality control plot comparing the amplification
status to the crt, Cq conf, and amplification score metrics output by
QuantStudio 12K Flex Software.
}
\examples{
data(example_se)
plot_qc(example_se)
}
