% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-visualizations.R
\name{plot_qc_from_se}
\alias{plot_qc_from_se}
\title{Plot a 3D Quality Control Graphic from a SummarizedExperiment}
\usage{
plot_qc_from_se(se)
}
\arguments{
\item{se}{a SummarizedExperiment object containing OpenArray qPCR data}
}
\value{
a plotly figure
}
\description{
Generates a 3-dimensional quality control plot comparing the amplification
status to the crt, Cq conf, and amplification score metrics output by
QuantStudio 12K Flex Software.
}
