% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{plot_raster}
\alias{plot_raster}
\title{Basic raster plot}
\usage{
plot_raster(r, title, ncolors = 1000, color_palette = c("#5E85B8",
  "#EDF0C0", "#C13127"))
}
\arguments{
\item{color_palette}{}
}
\description{
Basic raster plot
}
