% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCrt.R
\name{plotCrt}
\alias{plotCrt}
\title{Plot Relative Cycle Threshold Values by Gene}
\usage{
plotCrt(se)
}
\arguments{
\item{se}{a SummarizedExperiment object containing OpenArray qPCR data}
}
\value{
a ggplot2 figure
}
\description{
Generates a box and whisker plot visualizing the distribution of Crt values
measured on an OpenArray plate by Gene.
}
\examples{
data(example_se)

plotCrt(example_se)
}
