% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOverview.R
\name{plotOverview}
\alias{plotOverview}
\title{Plot Overview of Amplifications as a Heat Map}
\usage{
plotOverview(se)
}
\arguments{
\item{se}{a SummarizedExperiment object containing OpenArray qPCR data}
}
\value{
a ggplot2 figure
}
\description{
This function generates a heat map summarizing the amplification status
results for each sample on the OpenArray plate by target gene.
}
\examples{
data(example_se)
plotOverview(example_se)
}
