% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_data.R
\name{result_data}
\alias{result_data}
\title{Result data}
\format{
An R tibble (\code{.rda}) with twenty-two columns:
\describe{
\item{well}{integer value representing the well number of the reaction}
\item{well_position}{factor value representing the well corresponding to the reaction}
\item{sample_name}{factor storing name of the sample}
\item{target_name}{factor storing name of the assay target}
\item{crt}{numeric value for relative cycle threshold output by QS 12K Flex Software}
\item{amp_score}{numeric value for amplification score output by QS 12K Flex Software}
\item{cq_conf}{numeric value for confidence in Cq call output by QS 12K Flex Software}
\item{amp_status}{factor storing run outcome as called by QS 12K Flex Software}
\item{cycle}{integer value for cycle number}
\item{fam}{numeric value for multicomponent fluorescence recorded at the given cycle number}
\item{batch_name}{factor storing run name associated with the plate}
\item{regression_type}{character string denoting type of regression computed for the reaction}
\item{fam_pred}{numeric value for fluorescence value predicted by model}
\item{x_mid}{numeric cycle point at which the model curve crosses the midpoint between its minimal and maximal values; set to 40 for linear models}
\item{slope}{numeric slope of the model at the x_mid cycle point}
\item{delta}{numeric value approximating total change in fluorescence over the reaction}
\item{slope_threshold}{numeric value indicating minimum expected slope for amplification}
\item{delta_threshold}{numeric value indicating minimum expected overall change-in-fluorescence for amplification}
\item{crt_threshold}{numeric value indicating maximum expected crt for amplification}
\item{slope_pass}{logical expressing whether the reaction passed the slope condition}
\item{delta_pass}{logical expressing whether the reaction passed the change-in-fluorescence condition}
\item{crt_pass}{logical expressing whether the reaction passed the crt condition}
\item{result}{character indicating pass/fail for reaction amplification}
}
}
\source{
Generated by calling the assign_calls_with_key() function on the example curve_fit_data using the
example target-threshold key
}
\description{
This R tibble (.rda) contains qPCR, model curve, and result data in tidyverse-friendly format.
}
\details{
Example R objects are stored in the \verb{data/} directory of the package. Access this file with:

\code{result_data}
}
\keyword{data}
