% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-pipeline_core.R
\name{run_fit_curve}
\alias{run_fit_curve}
\title{Run curve-fitting function}
\usage{
run_fit_curve(data, linear_threshold)
}
\arguments{
\item{data}{A tibble of multicomponent fluorescence data vs. cycle number for one reaction}

\item{linear_threshold}{An integer describing the maximum overall change in fluorescence at which fluroescence curves will trivially be fitted to a linear regression}
}
\value{
A list object describing the model
}
\description{
An R wrapper for the internal python function, fit_curve(), which attempts to fit two-dimensional
multicomponent fluorescence vs. cycle number data to a 5-parameter logistic regression. For reactions
with a total change in fluorescence below the linear_threshold, returns instead
}
\examples{
data(tidy_run_data)
data <- dplyr::filter(tidy_run_data, well == 2794)
curve <- run_fit_curve(data, linear_threshold = 400)
}
