% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFitCurve.R
\name{runFitCurve}
\alias{runFitCurve}
\title{Run curve-fitting function}
\usage{
runFitCurve(data, linear_threshold)
}
\arguments{
\item{data}{A tibble of multicomponent fluorescence data vs.
cycle number for one reaction}

\item{linear_threshold}{An integer describing the maximum overall
change in fluorescence at which fluorescence curves will be trivially
fit to a linear regression}
}
\value{
A list object describing the model
}
\description{
An R wrapper for the internal python function, fit_curve(), which attempts
to fit two-dimensional multicomponent fluorescence vs. cycle number data to
a 5-parameter logistic regression. For reactions with a total change in
fluorescence below the linear_threshold, returns instead a linear regression
}
\examples{
data(example_se)

df <- as.data.frame(SummarizedExperiment::assay(example_se)) |> 
    dplyr::select(well_2385)

model <- data.frame(
    cycle = as.integer(sub("cycle_", "", rownames(df))),
    fam = as.numeric(df$well_2385)
)
}
