% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_to_qPCRBatch.R
\name{se_to_qPCRBatch}
\alias{se_to_qPCRBatch}
\title{Convert to qPCRBatch Object}
\usage{
se_to_qPCRBatch(se)
}
\arguments{
\item{se}{A SummarizedExperiment object with OpenArray run data}
}
\value{
a qPCRBatch object
}
\description{
Transforms OpenArray run data contained within the SummarizedExperiment
container into a qPCRBatch object. This conversion allows for convenient
gene expression analyses with the NormqPCR package.
}
\examples{
path = system.file(
    "extdata", 
    "oa_gene_expression_batch1.xlsx", 
    package = "OAtools"
)

se <- excel_to_summarized_experiment(path = path, num_results = 96)

qpcr <- se_to_qPCRBatch(se)
}
