% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seToQPCRBatch.R
\name{seToQPCRBatch}
\alias{seToQPCRBatch}
\title{Convert to qPCRBatch Object}
\usage{
seToQPCRBatch(se)
}
\arguments{
\item{se}{A SummarizedExperiment object with OpenArray run data}
}
\value{
a qPCRBatch object
}
\description{
Transforms OpenArray run data contained within the SummarizedExperiment
container into a qPCRBatch object. This conversion allows for convenient
gene expression analyses with the NormqPCR package.
}
\examples{
path = system.file(
    "extdata", 
    "oa_gene_expression_1.xlsx", 
    package = "OAtools"
)

se <- excelToSE(excel_path = path)

qpcr <- seToQPCRBatch(se)
}
