% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-data_import.R
\name{tidy_cumulative_gene_expression_data}
\alias{tidy_cumulative_gene_expression_data}
\title{Tidy cumulative gene expression data}
\usage{
tidy_cumulative_gene_expression_data(directory, num_results = 2688)
}
\arguments{
\item{directory}{file path to run data directory, which may hold several .xlsx files exported from QuantStudio 12K Flex Software}

\item{num_results}{an integer reflecting the number of rows on each 'results' sheet}
}
\value{
A tibble
}
\description{
This function transforms bulk raw gene expression run data exported in .xlsx format from
the OpenArray QuantStudio 12K Flex qPCR platform into a tidy R tibble.
}
\examples{
dir = system.file("extdata", package = "OAtools")
cumulative_run_data <- tidy_cumulative_gene_expression_data(directory = dir, num_results = 96)
}
