% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-tidy_run_data.R
\name{tidy_run_data}
\alias{tidy_run_data}
\title{Tidy run data}
\format{
An R tibble (\code{.rda}) with eleven columns:
\describe{
\item{well}{integer value representing the well number of the reaction}
\item{well_position}{factor value representing the well
corresponding to the reaction}
\item{sample_name}{factor storing name of the sample}
\item{target_name}{factor storing name of the assay target}
\item{crt}{numeric value for relative cycle threshold output by
QS 12K Flex Software}
\item{amp_score}{numeric value for amplification score output by
QS 12K Flex Software}
\item{cq_conf}{numeric value for confidence in Cq call output by
QS 12K Flex Software}
\item{amp_status}{factor storing run outcome as called by
QS 12K Flex Software}
\item{cycle}{integer value for cycle number}
\item{fam}{numeric value for multicomponent fluorescence recorded
at the given cycle number}
\item{batch_name}{factor storing run name associated with the plate}
}
}
\source{
Generated by calling the tidy_gene_expression_data() function on
the example raw data 'oa_gene_expression_batch1.xlsx'
}
\description{
This R tibble (.rda) contains imported run data in
tidyverse-friendly format.
}
\details{
Example R objects are stored in the \verb{data/} directory of the package.
Access this file with:

\code{tidy_run_data}
}
\keyword{data}
