% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildApp.R
\name{buildApp}
\alias{buildApp}
\title{A Shiny App for OAtools}
\usage{
buildApp()
}
\value{
A Shiny web application
}
\description{
Builds a Shiny application for interactively running OAtools
}
\examples{
app <- buildApp()
}
