% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determinePCRResults.R
\name{determinePCRResults}
\alias{determinePCRResults}
\title{Determine PCR results from features of multicomponent fluorescence curves}
\usage{
determinePCRResults(se, key_path)
}
\arguments{
\item{se}{a SummarizedExperiment object containing OpenArray qPCR Data}

\item{key_path}{file path to the target-threshold key}
}
\value{
a SummarizedExperiment
}
\description{
Assigns positive or negative PCR results to every PCR reaction depending on
the equation of the model curve optimized to the fluorescence curve. In
particular, the reporter dye fluorescence without normalization is used to
distinguish between real amplification and false positives.
}
\details{
The target-threshold key is an Excel file which stores threshold values
used to separate PCR curves. These thresholds are defined across metrics
such as cycle threshold, overall change-in-fluorescence, and slope of the
fluorescence curve during the exponential phase. The thresholds are
separately defined for each gene to accommodate inherent differences in
assays.
}
\examples{
data(example_se)

key_path = system.file(
    "extdata", 
    "target_threshold_key.xlsx", 
    package = "OAtools"
)

se <- example_se |> 
    determinePCRResults(key_path = key_path)
}
