% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeModels.R
\name{.runFitCurve}
\alias{.runFitCurve}
\title{Run the fit_curve() function}
\usage{
.runFitCurve(data, linear_threshold)
}
\arguments{
\item{data}{A \code{data.frame} with the following required columns:
\describe{
\item{cycle}{PCR cycle number}
\item{fluo}{observed fluorescence}
}}

\item{linear_threshold}{numeric value specifying the minimum overall
change-in-fluorescence over the PCR reaction required for the optimizer
to attempt fitting a logistic model}
}
\value{
The model as a nested list
}
\description{
A thin R wrapper for the python3 function fit_curve(), which attempts to
optimize 5-parameter logistic regressions to PCR fluorescence curves.
}
