% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildReporting.R
\name{.serverReporting}
\alias{.serverReporting}
\title{Reporting Server}
\usage{
.serverReporting(id, se)
}
\arguments{
\item{id}{identifier unique to \code{Reporting} page}

\item{se}{SummarizedExperiment with OpenArray assay data}
}
\value{
The \code{Reporting} server component
}
\description{
Reporting Server
}
