% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_se.R
\docType{data}
\name{example_se}
\alias{example_se}
\title{Example OpenArray Gene Expression Data Contained in a SummarizedExperiment}
\format{
a SummarizedExperiment object with the following:
\describe{
\item{colData}{information associated with each PCR well}
\item{rowData}{cycle numbers}
\item{assays}{matrices of fluorescence values by PCR well and cycle}
}
}
\source{
generated by running the following commands:

\code{filename <- "oa_gene_expression_1.xlsx"}

\code{path <- system.file("extdata", filename, package = "OAtools")}

\code{example_se <- excelToSE(excel_path = path)}

\code{example_se <- example_se |> fitModelsToSE(linear_threshold = 400)}
}
\description{
A sample of OpenArray gene expression data from respiratory tract microbiota
profiling experiments conducted at the UW Virology research lab on human
nasal swabs. This file (.rda) stores the PCR data in a SummarizedExperiment
container and is intended for use with package examples and unit testing.
}
\details{
Load this object into the environment with: \code{data(example_se)}
}
\keyword{data}
