% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_se.R
\docType{data}
\name{example_se}
\alias{example_se}
\title{OA qPCR Data from a Gene Expression Experiment}
\format{
a SummarizedExperiment object with the following:
\describe{
\item{colData}{metadata for each through-hole on the OpenArray Plate}
\item{rowData}{cycles over the PCR reaction}
\item{assay}{a matrix of fluorescence data for each through-hole on the
OpenArray Plate over the PCR reaction}
}
}
\source{
generated by running the excelToSummarizedExperiment() and
fitModelsToSe() functions on excel data exported from QuantStudio 12K Flex
software following a gene expression experiment.
}
\description{
This example data (.rda) contains a sample of OpenArray gene expression data
from a respiratory pathogen screening experiment conducted at the UW
Virology lab on human nasal swabs. This file is intended for use with
package examples, vignettes, and unit testing. It contains the data within
an instance of the Bioconductor S4 class SummarizedExperiment.
}
\details{
This data is stored in the 'data/' directory of the package.
Load into the environment with: \code{data(example_se)}
}
\keyword{data}
