% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{generate_report}
\alias{generate_report}
\title{Generate a report}
\usage{
generate_report(data)
}
\arguments{
\item{data}{a tibble of qPCR data ready for reporting}
}
\value{
an .html report summarizing the OpenArray run results
}
\description{
Generates a .html report
}
\examples{
generate_report(result_data)
}
