% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{generate_report}
\alias{generate_report}
\title{Generate a report}
\usage{
generate_report(data, path, analysis = "native")
}
\arguments{
\item{data}{a tibble of qPCR data ready for reporting}

\item{path}{intended outfile path}

\item{analysis}{character string for analytic method, supports 'native' and 'curve-fitting' approaches}
}
\value{
an .html report summarizing the OpenArray run results
}
\description{
Generates a .html report
}
\examples{
path = file.path("..", "reports")
generate_report(result_data, path)
}
