% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateReport.R
\name{generateReport}
\alias{generateReport}
\title{Generate a report}
\usage{
generateReport(se, path = tempdir())
}
\arguments{
\item{se}{a SummarizedExperiment containing OpenArray qPCR data}

\item{path}{intended outfile path, defaults to a temporary directory}
}
\value{
an .html report summarizing the OpenArray run results
}
\description{
Generates a .html report summarizing the OpenArray experiment.
}
\examples{
data(example_se)
generateReport(se = example_se)
}
