% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{lyr_to_tif}
\alias{lyr_to_tif}
\title{Get Raster Layer from sdmpredictors}
\usage{
lyr_to_tif(lyr, tif, crs, dir_sdm_cache, res = 10000, extent_crop = NULL,
  method = "ngb", redo_tif = FALSE)
}
\arguments{
\item{redo_tif}{}
}
\value{
raster object after loading from sdmpredictors, projecting and cropping to study area,
and writing raster. The function simply loads the existing raster if the tif already exists,
unless redo_tif = TRUE, in which case it is recreated with the arguments given.
}
\description{
Get Raster Layer from sdmpredictors
}
