% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_crt.R
\name{plot_crt}
\alias{plot_crt}
\title{Plot crt values}
\usage{
plot_crt(data, batch = NA)
}
\arguments{
\item{data}{a tibble of qPCR data; either a single batch or a cumulative tibble with the batch parameter specified}

\item{batch}{character string for input batch for plotting}
}
\value{
a ggplot2 figure
}
\description{
This function generates a box and whisker plot visualizing the distribution of
observed crt values for each assay target in a given run.
}
\examples{
fig <- plot_crt(tidy_run_data)
}
