% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_crt.R
\name{plot_crt}
\alias{plot_crt}
\title{Plot Crt Values from SummarizedExperiment}
\usage{
plot_crt(se)
}
\arguments{
\item{se}{a SummarizedExperiment object containing OpenArray qPCR data}
}
\value{
a ggplot2 figure
}
\description{
Generates a box and whisker plot to visualize the distribution of observed
Crt values for each assay target in an OpenArray plate.
}
\examples{
data(example_se)
plot_crt(example_se)
}
