% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-visualizations.R
\name{plot_overview}
\alias{plot_overview}
\title{Plot overview heat map}
\usage{
plot_overview(data, sample_list = NA, target_list = NA, batch = NA)
}
\arguments{
\item{data}{a tibble of qPCR data; either a single batch or a cumulative tibble with the batch parameter specified}

\item{sample_list}{optional character list of samples with which to complete heat map}

\item{target_list}{optional character list of assay targets with which to complete heat map}

\item{batch}{character string for input batch for plotting}
}
\value{
a ggplot2 figure
}
\description{
This function generates a heat map summarizing the amplification status results for
technical replicates grouped by sample, assay target, and batch name.
}
\examples{
fig <- plot_overview(tidy_run_data)
}
