% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-visualizations.R
\name{plot_qc}
\alias{plot_qc}
\title{Plot a 3-dimensional quality control graphic}
\usage{
plot_qc(data, batch = NA)
}
\arguments{
\item{data}{a tibble of qPCR data; either a single batch or a
cumulative tibble with the batch parameter specified}

\item{batch}{character string for input batch for plotting}
}
\value{
a plotly figure
}
\description{
This function generates a 3-dimensional QC plot comparing the
amplification status to the crt, Cq conf, and amplification score
output by QuantStudio 12K Flex software.
}
\examples{
data(tidy_run_data)
fig <- plot_qc(tidy_run_data)
}
