% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{plot_raster}
\alias{plot_raster}
\title{Basic raster plot}
\usage{
plot_raster(r, title, ncolors = 1000, color_palette = c("#5E85B8",
  "#EDF0C0", "#C13127"))
}
\arguments{
\item{r}{raster}

\item{title}{title of plot}

\item{ncolors}{number of colors (default: 1000)}

\item{color_palette}{color palette (default: c("#5E85B8","#EDF0C0","#C13127"))}
}
\value{
Function does not return any value, but plot is generated with the default output device.
}
\description{
Basic raster plot
}
