% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCrt.R
\name{plotCrt}
\alias{plotCrt}
\title{Plot Crt Values from SummarizedExperiment}
\usage{
plotCrt(se)
}
\arguments{
\item{se}{a SummarizedExperiment object containing OpenArray qPCR data}
}
\value{
a ggplot2 figure
}
\description{
Generates a box and whisker plot to visualize the distribution of observed
Crt values for each assay target in an OpenArray plate.
}
\examples{
data(example_se)
plotCrt(example_se)
}
