% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotModel.R
\name{plotModel}
\alias{plotModel}
\title{Plot Fluorescence Values Predicted by Model}
\usage{
plotModel(se, well_id)
}
\arguments{
\item{se}{a SummarizedExperiment object containing OpenArray qPCR data}

\item{well_id}{a character representing the name of the well to plot as
listed in the assay matrix}
}
\value{
a ggplot2 figure
}
\description{
Juxtaposes the fluorescence values predicted by the model optimized to the
measured fluorescence vs. cycle data for a particular well.
}
\examples{
data(example_se)

plotModel(example_se, well_id = "well_2665")
}
