% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOverview.R
\name{plotOverview}
\alias{plotOverview}
\title{Plot Amplification Status by Sample and Gene}
\usage{
plotOverview(se)
}
\arguments{
\item{se}{a SummarizedExperiment object containing OpenArray qPCR data}
}
\value{
a ggplot2 figure
}
\description{
Generates an overview graphic summarizing qPCR results for each combination
of sample and gene on an OpenArray experiment.
}
\examples{
data(example_se)

plotOverview(example_se)
}
